% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud4R.R
\docType{package}
\name{cloud4R}
\alias{cloud4R}
\alias{cloud4R-package}
\title{Tools to Harmonize (meta)data management on cloud repositories}
\description{
Provides tools to harmonize the (meta)data manageement on cloud repositories, using standard formats and protocols.
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
