% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CloudManager.R
\docType{class}
\name{CloudManager}
\alias{CloudManager}
\title{CloudManager}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an CloudManager
}
\description{
CloudManager

CloudManager
}
\note{
Main user class to be used with \pkg{cloud4R}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{cloud}
\keyword{manager}
\section{Super class}{
\code{\link[cloud4R:cloud4RLogger]{cloud4R::cloud4RLogger}} -> \code{CloudManager}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{systems}}{one or more cloud systems}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CloudManager$new()}}
\item \href{#method-clone}{\code{CloudManager$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="cloud4R" data-topic="cloud4RLogger" data-id="ERROR">}\href{../../cloud4R/html/cloud4RLogger.html#method-ERROR}{\code{cloud4R::cloud4RLogger$ERROR()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cloud4R" data-topic="cloud4RLogger" data-id="getClass">}\href{../../cloud4R/html/cloud4RLogger.html#method-getClass}{\code{cloud4R::cloud4RLogger$getClass()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cloud4R" data-topic="cloud4RLogger" data-id="getClassName">}\href{../../cloud4R/html/cloud4RLogger.html#method-getClassName}{\code{cloud4R::cloud4RLogger$getClassName()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cloud4R" data-topic="cloud4RLogger" data-id="INFO">}\href{../../cloud4R/html/cloud4RLogger.html#method-INFO}{\code{cloud4R::cloud4RLogger$INFO()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cloud4R" data-topic="cloud4RLogger" data-id="logger">}\href{../../cloud4R/html/cloud4RLogger.html#method-logger}{\code{cloud4R::cloud4RLogger$logger()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cloud4R" data-topic="cloud4RLogger" data-id="WARN">}\href{../../cloud4R/html/cloud4RLogger.html#method-WARN}{\code{cloud4R::cloud4RLogger$WARN()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize the cloud manager
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CloudManager$new(systems, logger = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{systems}}{one or more systems}

\item{\code{logger}}{logger}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CloudManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
