% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CloudSystem.R
\docType{class}
\name{CloudSystem}
\alias{CloudSystem}
\title{CloudSystem}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an CloudSystem
}
\description{
CloudSystem

CloudSystem
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{cloud}
\keyword{system}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{system identifier}

\item{\code{handler}}{system handler}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CloudSystem$new()}}
\item \href{#method-getId}{\code{CloudSystem$getId()}}
\item \href{#method-getHandler}{\code{CloudSystem$getHandler()}}
\item \href{#method-clone}{\code{CloudSystem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a cloud system
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CloudSystem$new(id, handler)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{cloud system identifier}

\item{\code{handler}}{cloud system function handler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getId"></a>}}
\if{latex}{\out{\hypertarget{method-getId}{}}}
\subsection{Method \code{getId()}}{
Get Id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CloudSystem$getId()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the system identifier
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getHandler"></a>}}
\if{latex}{\out{\hypertarget{method-getHandler}{}}}
\subsection{Method \code{getHandler()}}{
Get system handler
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CloudSystem$getHandler()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the system handler
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CloudSystem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
