% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CloudSystemInstance.R
\docType{class}
\name{CloudSystemInstance}
\alias{CloudSystemInstance}
\title{CloudSystemInstance}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an CloudSystemInstance
}
\description{
CloudSystemInstance

CloudSystemInstance
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{cloud}
\keyword{system}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{system identifier}

\item{\code{refid}}{system reference identifier}

\item{\code{instance}}{system instance}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CloudSystemInstance$new()}}
\item \href{#method-getId}{\code{CloudSystemInstance$getId()}}
\item \href{#method-getInstance}{\code{CloudSystemInstance$getInstance()}}
\item \href{#method-clone}{\code{CloudSystemInstance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a cloud system instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CloudSystemInstance$new(id, refid, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{instance identifier}

\item{\code{refid}}{cloud system reference identifier}

\item{\code{...}}{arguments to set for the system instance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getId"></a>}}
\if{latex}{\out{\hypertarget{method-getId}{}}}
\subsection{Method \code{getId()}}{
Get Id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CloudSystemInstance$getId()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the system instance identifier
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getInstance"></a>}}
\if{latex}{\out{\hypertarget{method-getInstance}{}}}
\subsection{Method \code{getInstance()}}{
Get system instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CloudSystemInstance$getInstance()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the system instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CloudSystemInstance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
