% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CloudSystem.R
\name{get_cloud_system}
\alias{get_cloud_system}
\title{get_cloud_system}
\usage{
get_cloud_system(id)
}
\arguments{
\item{id}{system id}
}
\value{
an object of class \link{CloudSystem}, \code{NULL} if not system identified
}
\description{
Get a cloud system given an ID
}
