% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CloudSystem.R
\name{list_cloud_systems}
\alias{list_cloud_systems}
\title{list_cloud_systems}
\usage{
list_cloud_systems()
}
\value{
the list of supported cloud systems
}
\description{
\code{list_cloud_systems} lists the available cloud systems currently managed in \pkg{cloud4R}
}
